[FEATURE] Enhanced event details rendering
The rendering of the event details (description, images, files, location,
organizer) has been moved to a new EventDetails partial.

The rendering of the description, images and files is now handled by
the default content rendering TypoScript for textpic or uploads content
elements.

For this to work the image rendering needs to be done by FAL (introduced
by https://review.typo3.org/25511/, not yet merged).