[TASK] Removed custom dispatcher
Since Extbase in the current version allows to override the available
actions by FlexForm configuration (switchableControllerActions) the
custom dispatcher code in the base controller was removed.

The FlexForm configuration was modified so that the integrator is able
to configure flexible controller / action select items for the
switchableControllerActions setting.

The custom view classes and the controller name properties and the have
been removed.