[FEATURE] Improved back link for list view
The link to the detail view appends the current query string so that
it is possible to link back to the original query in the detail view.

This is realized by setting the addQueryString parameter of the
link.action view helper to true in the list view and the detail view.