[!!!][TASK] Unified address handling
The address properties for the location and the organizer of an event
have been removed from the event domain model and have been replaced
with an inline address record.

This makes enhancements of the address records easier in the future.

The inline address records are hidden in the record list by a hook.