[!!!][FEATURE] Enhanced / restructured exception handling
The database model for the exceptions has changed and is now more
integrated into the Extbase data handling. Exceptions are now attached
as inline elements to events and exception groups to make the usage
easier for editors.

A new exception type is introduces that allows overwriting event
properties for the exceptional events. E.g. an event can now be
marked as "Cancelled" on a certain date within a recurring event series.