[!!!][FEATURE] Improved multi language handling
Only language relevant event fields are now editable in event
translations.

The exceptions are now translatable, this makes sense when event
properties like the teaser are overwritten.

For this to work properly the event index needs to be translatable
as well.

The event and event index slugs are also translatable now.

Warning! After applying this patch the event index needs to be rebuild!