[BUGFIX] Improve slug handling
The slug generation is now triggerd by the indexer. This makes sure
that missing slugs will be created during the scheduler run.

The slug generation is now consolidated in the generateSlug() method
in the Event class.

The makeSlugUnique() method in the EventIndexRepository is now
respecting the sys_language which allows identical slugs in
different languages.