[BUGFIX] Bring back action fallback functionality
Since the controller action configuration was restructured to
match the Extbase defaults the action fallback mechanism does
not work any more.

To resolve this the resolveActionMethodName() method is overwritten
and checks for a useAction configuration if the current action
method can not be resolved.

The fallback methods for the different actions are added to the
configuration.