[!!!] Remove magic methods in domain models
The workarounds for is() and has() methods in the domain models
are removed because in the current version Extbase allows
getter methods starting with "is".

To make use of this functionality we also need to remove the
magic function for retrieving event properties in the EventIndex
domain model because otherwise the is_callable detection of
the ObjectAccess utility will fail. The magic method is
replaced with actual getter methods in the EventIndex class.

A goal for the future is the complete removal of the event getters
in the EventIndex model.